refactored native type declarations
